kernel: move MarkXXXBags helpers to new bags.c file
