+++ src/bags.c
@@ -1,0 +1,51 @@
+/****************************************************************************
+**
+**/
+
+#include "gasman.h"
+
+inline void MarkArrayOfBags(const Bag array[], UInt count)
+{
+    for (UInt i = 0; i < count; i++) {
+        MarkBag(array[i]);
+    }
+}
+
+void MarkNoSubBags(Bag bag)
+{
+}
+
+void MarkOneSubBags(Bag bag)
+{
+    MarkArrayOfBags(CONST_PTR_BAG(bag), 1);
+}
+
+void MarkTwoSubBags(Bag bag)
+{
+    MarkArrayOfBags(CONST_PTR_BAG(bag), 2);
+}
+
+void MarkThreeSubBags(Bag bag)
+{
+    MarkArrayOfBags(CONST_PTR_BAG(bag), 3);
+}
+
+void MarkFourSubBags(Bag bag)
+{
+    MarkArrayOfBags(CONST_PTR_BAG(bag), 4);
+}
+
+void MarkAllSubBags(Bag bag)
+{
+    MarkArrayOfBags(CONST_PTR_BAG(bag), SIZE_BAG(bag) / sizeof(Bag));
+}
+
+void MarkAllButFirstSubBags(Bag bag)
+{
+    MarkArrayOfBags(CONST_PTR_BAG(bag) + 1, SIZE_BAG(bag) / sizeof(Bag) - 1);
+}
+
+void MarkAllSubBagsDefault(Bag bag)
+{
+    MarkArrayOfBags(CONST_PTR_BAG(bag), SIZE_BAG(bag) / sizeof(Bag));
+}
